+++ test/flop/schema_test.exs
@@ -7,6 +7,8 @@ defmodule Flop.SchemaTest do
   doctest Flop.Schema, import: true
 
   defmodule Panini do
+    use Ecto.Schema
+
     @derive {Flop.Schema,
              filterable: [:name, :age],
              sortable: [:name, :age, :topping_count],
@@ -17,7 +19,12 @@ defmodule Flop.SchemaTest do
                order_directions: [:desc, :asc]
              },
              compound_fields: [name_or_email: [:name, :email]],
-             join_fields: [topping_name: {:toppings, :name}],
+             join_fields: [
+               topping_name: [
+                 binding: :toppings,
+                 field: :name
+               ]
+             ],
              alias_fields: [:topping_count],
              custom_fields: [
                inserted_at: [
@@ -26,7 +33,11 @@ defmodule Flop.SchemaTest do
                ]
              ]}
 
-    defstruct [:name, :email, :age]
+    schema "paninis" do
+      field :name, :string
+      field :email, :string
+      field :age, :integer
+    end
   end
 
   test "default_order/1 returns the default order passed as an option" do
@@ -38,38 +49,6 @@ defmodule Flop.SchemaTest do
 
   test "default_limit/1 returns the default limit passed as option" do
     assert Schema.default_limit(%Panini{}) == 20
-  end
-
-  test "field_type/2 returns :normal tuple for normal fields" do
-    assert Schema.field_type(%Panini{}, :name) == {:normal, :name}
-    assert Schema.field_type(%Panini{}, :age) == {:normal, :age}
-  end
-
-  test "field_type/2 returns config for compound fields" do
-    assert Schema.field_type(%Panini{}, :name_or_email) ==
-             {:compound, [:name, :email]}
-  end
-
-  test "field_type/2 returns config for join fields" do
-    assert Schema.field_type(%Panini{}, :topping_name) ==
-             {:join,
-              %{
-                binding: :toppings,
-                field: :name,
-                path: [:toppings, :name],
-                ecto_type: nil
-              }}
-  end
-
-  test "field_type/2 returns config for custom fields" do
-    assert Schema.field_type(%Panini{}, :inserted_at) ==
-             {:custom,
-              %{
-                filter: {Panini, :date_filter, [some: "option"]},
-                ecto_type: :date,
-                operators: nil,
-                bindings: []
-              }}
   end
 
   test "max_limit/1 returns the max limit passed as option" do
@@ -85,12 +64,6 @@ defmodule Flop.SchemaTest do
   test "calling default_order/1 without deriving raises error" do
     assert_raise Protocol.UndefinedError, fn ->
       Schema.default_order(%{})
-    end
-  end
-
-  test "calling field_type/2 without deriving raises error" do
-    assert_raise Protocol.UndefinedError, fn ->
-      Schema.field_type(%{}, :field)
     end
   end
 
@@ -213,7 +186,12 @@ defmodule Flop.SchemaTest do
               Flop.Schema,
               filterable: [],
               sortable: [],
-              join_fields: [name: {:eater, :name}],
+              join_fields: [
+                name: [
+                  binding: :eater,
+                  field: :name
+                ]
+              ],
               compound_fields: [name: [:name, :nickname]]
             }
             defstruct [:name, :nickname]
@@ -249,7 +227,12 @@ defmodule Flop.SchemaTest do
               Flop.Schema,
               filterable: [],
               sortable: [],
-              join_fields: [owner_name: {:owner, :name}],
+              join_fields: [
+                owner_name: [
+                  binding: :owner,
+                  field: :name
+                ]
+              ],
               alias_fields: [:owner_name]
             }
             defstruct [:id]
@@ -289,7 +272,12 @@ defmodule Flop.SchemaTest do
               Flop.Schema,
               filterable: [],
               sortable: [],
-              join_fields: [owner_name: {:owner, :name}],
+              join_fields: [
+                owner_name: [
+                  binding: :owner,
+                  field: :name
+                ]
+              ],
               custom_fields: [
                 owner_name: [
                   filter: {__MODULE__, :some_function, []}